git-svn-id: https://svn.code.sf.net/p/freedos-32/code/trunk@75 85e9c19b-ca11-0410-bd92-afdf7feba87b

+++ fd32/filesys/fs.c
@@ -238,7 +238,7 @@ int fd32_unlink(char *FileName, DWORD Flags)
 /* The LSEEK system call.                                     */
 /* Calls the "lseek" function of the device hosting the file. */
 /* Returns 0 on success, or a negative error code on failure. */
-int fd32_lseek(int Handle, LONGLONG Offset, int Origin, LONGLONG *Result)
+int fd32_lseek(int Handle, long long int Offset, int Origin, long long int *Result)
 {
   fd32_lseek_t  S;
   tJft         *J;
